requirements.txt
absl-py==0.8.0
astor==0.8.0
boto==2.49.0
boto3==1.9.220
botocore==1.12.220
certifi==2019.6.16
chardet==3.0.4
docutils==0.15.2
gast==0.2.2
gensim==3.8.0
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
jmespath==0.9.4
joblib==0.13.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mecab-python-windows==0.996.3
numpy==1.17.1
protobuf==3.9.1
pydot==1.4.1
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
smart-open==1.8.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
